requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
openai
slowapi
supabase
pyjwt
python-jose[cryptography]
